backend/requirements.txt
# Main Dependencies
fastapi>=0.115.2,<0.116.0
uvicorn==0.22.0
pydantic[email]==2.9.2
requests>=2.22.0,<3.0.0
loguru>=0.6.0,<0.7.0
joblib>=1.2.0,<2.0.0
certifi>=2022.12.7,<2023.0.0
python-dotenv>=1.0.0,<2.0.0
python-multipart>=0.0.6,<0.1.0
openai[embeddings]>=1.3.5,<2.0.0
aiohttp>=3.8.3,<4.0.0
fastapi-utils>=0.7.0,<0.8.0
envyaml>=1.10.211231,<2.0.0
tokenizers==0.20.3
pydub>=0.25.1,<0.26.0
llama_index==0.11.18
nest_asyncio
httpx
sqlalchemy
databases
sqlite-utils
pinecone-client>=5.0.1,<6.0.0
# langchain>=0.0.312,<0.0.313
nltk>=3.8.1,<4.0.0
PyMuPDF>=1.23.22,<2.0.0
llama-index-vector-stores-pinecone==0.2.1
google-generativeai==0.5.2
llama-index-embeddings-gemini==0.2.1
PyJWT>=2.7.0,<3.0.0
# Development Dependencies
pytest>=7.2,<8.0.0
black>=19.10b0,<20.0.0
autopep8>=2.0.0,<3.0.0
ipdb>=0.13.0,<0.14.0
pytest-cov==2.12.1
mypy==0.900
click==8.0.2
pre-commit>=3.0.0,<4.0.0
jupyterlab>=3.5.3,<4.0.0
passlib==1.7.4
bcrypt==4.2.1
python-jose == 3.3.0
autogen==0.3.0
langchain-community==0.3.8
langchain-openai==0.2.9